edit code

+++ application/controllers/UserController.php
@@ -1,13 +1,16 @@
 <?php
 if ( ! defined('BASEPATH')) exit('No direct script access allowed');
-class UserController extends CI_Controller{
-	function __construct(){
+class UserController extends CI_Controller
+{
+	function __construct()
+	{
 		parent::__construct();
 		$this->load->model('user');
 		//$this->output->cache(20);
 	}
 
-	public function createUser(){
+	public function createUser()
+	{
 		if($this->input->post()){
 			$this->form_validation->set_rules('username', 'username', 'required|min_length[6]|max_length[10]');
 			$this->form_validation->set_rules('password', 'password', 'required|min_length[6]|max_length[10]');
@@ -26,8 +29,11 @@ class UserController extends CI_Controller{
 		return $this->load->view('user/create');
 	}
 
-	public function updateUser(){
+	public function updateUser()
+	{
 		$id = $this->uri->segment(2);
+		if(!$this->user->findOrFail($id))
+			return $this->load->view('errors/error_404');
 		$info = $this->user->getUser($id);
 		if($this->input->post()){
 			$this->form_validation->set_rules('username','username', 'required|min_length[6]|max_length[10]');
@@ -48,14 +54,18 @@ class UserController extends CI_Controller{
 		return $this->load->view('user/edit',$data);
 	}
 
-	public function deleteUser(){
+	public function deleteUser()
+	{
 		$id = $this->uri->segment(2);
+		if(!$this->user->findOrFail($id))
+			return $this->load->view('errors/error_404');
 		$this->user->delete($id);
 		$this->session->set_flashdata('mess', 'delete success!');
 		return redirect(base_url());
 	}
 
-	public function login(){
+	public function login()
+	{
 		if($this->input->post()){
 			$this->form_validation->set_rules('username', 'username', 'required');
 			$this->form_validation->set_rules('password', 'password', 'required');
@@ -77,7 +87,8 @@ class UserController extends CI_Controller{
 		return $this->load->view('user/login');
 	}
 
-	public function index(){
+	public function index()
+	{
 		if(isset($_SESSION['user'])){
 			$users = $this->user->list();
 			$data['users'] = $users;
@@ -86,7 +97,8 @@ class UserController extends CI_Controller{
 		return redirect(base_url('login'));
 	}
 
-	public function logout(){
+	public function logout()
+	{
 		$this->session->sess_destroy();
 		return redirect(base_url('login'));
 	}
